Refactored BashpileMainTest hierarchy
