+++ src/test/java/com/bashpile/maintests/CommandSubstutionBashpileTest.java
@@ -1,7 +1,6 @@
-package com.bashpile;
+package com.bashpile.maintests;
 
 import com.bashpile.commandline.ExecutionResults;
-import com.bashpile.testhelper.BashpileMainTest;
 import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
@@ -11,7 +10,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 @Order(60)
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-public class CommandSubstutionBashpileMainTest extends BashpileMainTest {
+public class CommandSubstutionBashpileTest extends BashpileTest {
 
     /** Simple one word command */
     @Test @Order(10)
