+++ src/test/java/com/bashpile/maintests/LexerBashpileTest.java
@@ -1,7 +1,6 @@
-package com.bashpile;
+package com.bashpile.maintests;
 
 import com.bashpile.commandline.ExecutionResults;
-import com.bashpile.testhelper.BashpileMainTest;
 import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
@@ -20,7 +19,7 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
  */
 @Order(10)
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-class LexerBashpileMainTest extends BashpileMainTest {
+class LexerBashpileTest extends BashpileTest {
 
     @Test
     @Order(10)
